build-requirements.txt
# in the meantime we migrate to pyproject.toml
# this represent the minimum requirement to build (for the optional speedup)
mypy==1.5.1; python_version >= '3.8'
mypy==1.4.1; python_version < '3.8'
build==0.10.0
wheel==0.41.1
